requirements.txt
cx-Oracle==8.3.0
py-logger @ git+https://github.com/jrenyang/py-logger.git
python-dotenv==0.19.2
